requirements.txt
aiobotocore==2.13.1

aiohttp==3.9.5

aioitertools==0.11.0

aiosignal==1.3.1

altair==5.3.0

annotated-types==0.7.0

antlr4-python3-runtime==4.9.3

anyio==4.4.0

appdirs==1.4.4

appnope==0.1.4

asgiref==3.8.1

asttokens==2.4.1

attrs==23.2.0

backoff==2.2.1

bcrypt==4.1.3

beautifulsoup4==4.12.3

blinker==1.8.2

botocore==1.34.131

build==1.2.1

cachelib==0.13.0

cachetools==5.3.3

certifi==2024.6.2

cffi==1.16.0

chardet==5.2.0

charset-normalizer==3.3.2

chroma-hnswlib==0.7.3

chromadb==0.5.3

click==8.1.7

colorama==0.4.6

coloredlogs==15.0.1

comm==0.2.2

contourpy==1.2.1

cryptography==42.0.8

cycler==0.12.1

dataclasses-json==0.6.7

datasets==2.20.0

debugpy==1.8.1

decorator==5.1.1

deepdiff==7.0.1

Deprecated==1.2.14

dill==0.3.8

distro==1.9.0

dnspython==2.6.1

docstring_parser==0.16

effdet==0.4.1

email_validator==2.2.0

emoji==2.12.1

et-xmlfile==1.1.0

executing==2.0.1

fastapi==0.111.0

fastapi-cli==0.0.4

filelock==3.15.4

filetype==1.2.0

Flask==3.0.3

Flask-Session==0.8.0

flatbuffers==24.3.25

fonttools==4.53.0

frozenlist==1.4.1

fsspec==2024.5.0

gcsfs==2024.6.1

gitdb==4.0.11

GitPython==3.1.43

google-api-core==2.19.1

google-auth==2.30.0

google-auth-oauthlib==1.2.0

google-cloud-aiplatform==1.57.0

google-cloud-bigquery==3.25.0

google-cloud-core==2.4.1

google-cloud-resource-manager==1.12.3

google-cloud-storage==2.17.0

google-cloud-vision==3.7.2

google-crc32c==1.5.0

google-resumable-media==2.7.1

googleapis-common-protos==1.63.2

greenlet==3.0.3

grpc-google-iam-v1==0.13.1

grpcio==1.64.1

grpcio-status==1.62.2

h11==0.14.0

httpcore==1.0.5

httptools==0.6.1

httpx==0.27.0

huggingface-hub==0.23.4

humanfriendly==10.0

idna==3.7

importlib_metadata==7.1.0

importlib_resources==6.4.0

intel-openmp==2021.4.0

iopath==0.1.10

ipykernel==6.29.4

ipython==8.25.0

itsdangerous==2.2.0

jedi==0.19.1

Jinja2==3.1.4

jmespath==1.0.1

joblib==1.4.2

jsonpatch==1.33

jsonpath-python==1.0.6

jsonpointer==3.0.0

jsonschema==4.22.0

jsonschema-specifications==2023.12.1

jupyter_client==8.6.2

jupyter_core==5.7.2

kiwisolver==1.4.5

kubernetes==30.1.0

langchain==0.2.6

langchain-chroma==0.1.2

langchain-community==0.2.6

langchain-core==0.2.10

langchain-google-vertexai==1.0.6

langchain-openai==0.1.13

langchain-text-splitters==0.2.2

langdetect==1.0.9

langsmith==0.1.82

layoutparser==0.3.4

lxml==5.2.2

Markdown==3.6

markdown-it-py==3.0.0

MarkupSafe==2.1.5

marshmallow==3.21.3

matplotlib==3.9.0

matplotlib-inline==0.1.7

mdurl==0.1.2

mkl==2021.4.0

mmh3==4.1.0

monotonic==1.6

mpmath==1.3.0

msgspec==0.18.6

multidict==6.0.5

multiprocess==0.70.16

mypy-extensions==1.0.0

nest-asyncio==1.6.0

networkx==3.3

nltk==3.8.1

numpy==1.26.4

oauthlib==3.2.2

olefile==0.47

omegaconf==2.3.0

onnx==1.16.1

onnxruntime==1.18.0

openai==1.35.7

opencv-python==4.10.0.84

openpyxl==3.1.4

opentelemetry-api==1.25.0

opentelemetry-exporter-otlp-proto-common==1.25.0

opentelemetry-exporter-otlp-proto-grpc==1.25.0

opentelemetry-instrumentation==0.46b0

opentelemetry-instrumentation-asgi==0.46b0

opentelemetry-instrumentation-fastapi==0.46b0

opentelemetry-proto==1.25.0

opentelemetry-sdk==1.25.0

opentelemetry-semantic-conventions==0.46b0

opentelemetry-util-http==0.46b0

ordered-set==4.1.0

orjson==3.10.5

overrides==7.7.0

packaging==24.1

pandas==2.2.2

parso==0.8.4

pdf2image==1.17.0

pdfminer==20191125

pdfminer.six==20231228

pdfplumber==0.11.1

pexpect==4.9.0

pikepdf==9.0.0

pillow==10.3.0

pillow_heif==0.16.0

pinecone==4.0.0

platformdirs==4.2.2

portalocker==2.10.0

posthog==3.5.0

prompt_toolkit==3.0.47

proto-plus==1.24.0

protobuf==4.25.3

psutil==5.9.8

ptyprocess==0.7.0

pure-eval==0.2.2

pyarrow==16.1.0

pyarrow-hotfix==0.6

pyasn1==0.6.0

pyasn1_modules==0.4.0

pycocotools==2.0.8

pycparser==2.22

pycryptodome==3.20.0

pydantic==2.7.4

pydantic_core==2.18.4

pydeck==0.9.1

Pygments==2.18.0

pypandoc==1.13

pyparsing==3.1.2

pypdf==4.2.0

pypdfium2==4.30.0

PyPika==0.48.9

pyproject_hooks==1.1.0

pyreadline3==3.4.1

pysbd==0.3.4

pytesseract==0.3.10

python-dateutil==2.9.0.post0

python-docx==1.1.2

python-dotenv==1.0.1

python-iso639==2024.4.27

python-magic==0.4.27

python-multipart==0.0.9

python-oxmsg==0.0.1

python-pptx==0.6.23

pytz==2024.1

pywin32==306

PyYAML==6.0.1

pyzmq==26.0.3

ragas==0.1.10

rapidfuzz==3.9.3

referencing==0.35.1

regex==2024.5.15

requests==2.32.3

requests-oauthlib==2.0.0

requests-toolbelt==1.0.0

rich==13.7.1

rpds-py==0.18.1

rsa==4.9

s3fs==2024.6.1

safetensors==0.4.3

scikit-learn==1.5.0

scipy==1.14.0

sentence-transformers==3.0.1

shapely==2.0.4

shellingham==1.5.4

six==1.16.0

smmap==5.0.1

sniffio==1.3.1

soupsieve==2.5

SQLAlchemy==2.0.31

st-files-connection==0.1.0

stack-data==0.6.3

starlette==0.37.2

streamlit==1.36.0

sympy==1.12.1

tabulate==0.9.0

tbb==2021.13.0

tenacity==8.4.2

threadpoolctl==3.5.0

tiktoken==0.7.0

timm==1.0.7

tokenizers==0.19.1

toml==0.10.2

toolz==0.12.1

torch==2.3.1

torchvision==0.18.1

tornado==6.4.1

tqdm==4.66.4

traitlets==5.14.3

transformers==4.42.1

typer==0.12.3

typing-inspect==0.9.0

typing_extensions==4.12.2

tzdata==2024.1

ujson==5.10.0

unstructured==0.14.8

unstructured-client==0.23.7

unstructured-inference==0.7.36

unstructured.pytesseract==0.3.12

urllib3==2.2.2

uvicorn==0.30.1

vertexai==1.43.0

watchdog==4.0.1

watchfiles==0.22.0

wcwidth==0.2.13

websocket-client==1.8.0

websockets==12.0

Werkzeug==3.0.3

wrapt==1.16.0

xlrd==2.0.1

XlsxWriter==3.2.0

xxhash==3.4.1

yarl==1.9.4

zipp==3.19.2

